🚚 rename: 패키지 경로 수정
