+++ src/main/java/kr/co/isajjim/global/sse/SseProvider.java
@@ -1,6 +1,5 @@
 package kr.co.isajjim.global.sse;
 
-import kr.co.isajjim.domains.estimate.persistence.repository.EmitterRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
